Make launchxl ccfg a separate binary
We don't want to flash CCFG data for the launchxl as part of the kernel.
It's dangerous, it makes the flah binary huge, and it makes it harder to
make board-specific changes.

Instead, pulls the ccfg into a separately compiled binary for the
launchxl board with it's own Makefile target to flash it to the ccfg
section. Other boards should do other board-specific things for CCFG
(including potentially leaving it untouched).